+++ lib/model/app/contributors.dart
@@ -169,6 +169,7 @@ class Contributors {
     Contributor('Víctor Humet Salado', 'ressaguer@proton.me'),
     Contributor('traaduttore', 'traaduttore@users.noreply.hosted.weblate.org'),
     Contributor('sgchr', 'pr3dali3n@gmail.com'),
+    Contributor('Zhafran Galeel', 'zhafgaleel@gmail.com'),
     // Contributor('Femini', 'nizamismidov4@gmail.com'), // Azerbaijani
     // Contributor('Jamil Farajov', 'jamilfarajov@gmail.com'), // Azerbaijani
     // Contributor('Alvi Khan', 'aveenalvi@gmail.com'), // Bengali