+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.14.3"></a>[v1.14.3] - 2026-04-06
+
+### Fixed
+
+- undated item handling for date editing
+
 ## <a id="v1.14.2"></a>[v1.14.2] - 2026-03-30
 
 ### Added